skip set local test
